Move room-related objects to its own module & also delete room when no player left in custom room
